Setting defuault value for Nspecies to 1
Setting defuault value for Nspecies to 1: this is necessary for cases
whem Voigt is used without Species specified. I would need to keep an
eye as it has not been tested.
Another change: enermax is removed from the energy section and added to
the lifetimes. Now it shoud give -1 for all states above enermax